Creating ManageConselors style folder and ManageAcceptedRegisters style file
