+++ src/Styles/ManageConselors/ManageAcceptedRegistersStyles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+const buttonBox = StyleSheet.create({
+    design: {
+      borderColor: 'black',
+      borderWidth: 1,
+      borderRadius: 7,
+      backgroundColor: '#FF3B30',
+      padding: 8,
+      justifyContent: 'center',
+      marginRight: 13,
+    },
+    text: {
+      fontSize: 12,
+      textAlign: 'center',
+    },
+});
+
+export default buttonBox
